pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xge"
version = "0.1.0"
description = "Crypto arbitrage detection across exchanges"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0.0",
    "redis>=5.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]
